+++ src/server/game/Guilds/Guild.h
@@ -216,7 +216,7 @@ struct GuildBankEventLogEntry
     uint8  EventType;
     uint32 PlayerGuid;
     uint32 ItemOrMoney;
-    uint8  ItemStackCount;
+    uint16  ItemStackCount;
     uint8  DestTabId;
     uint64 TimeStamp;
 
@@ -428,7 +428,7 @@ class Guild
         // Guild Bank Event Logs
         void   LoadGuildBankEventLogFromDB();
         void   DisplayGuildBankLogs(WorldSession *session, uint8 TabId);
-        void   LogBankEvent(SQLTransaction& trans, uint8 EventType, uint8 TabId, uint32 PlayerGuidLow, uint32 ItemOrMoney, uint8 ItemStackCount=0, uint8 DestTabId=0);
+        void   LogBankEvent(SQLTransaction& trans, uint8 EventType, uint8 TabId, uint32 PlayerGuidLow, uint32 ItemOrMoney, uint16 ItemStackCount=0, uint8 DestTabId=0);
         bool   AddGBankItemToDB(uint32 GuildId, uint32 BankTab , uint32 BankTabSlot , uint32 GUIDLow, uint32 Entry, SQLTransaction& trans);
 
     protected:
